requirements.txt
psycopg2==2.8
pymongo==3.6.1
PyYAML==3.12
